Make the specs happy
